# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Records API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/records": {
            "get": {
                "tags": [
                    "Records"
                ],
                "summary": "Retrieve all records",
                "operationId": "getAllRecords",
                "responses": {
                    "200": {
                        "description": "List of records",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Record"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Records"
                ],
                "summary": "Create a new record",
                "operationId": "createRecord",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Record"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Record created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/records/{id}": {
            "get": {
                "tags": [
                    "Records"
                ],
                "summary": "Retrieve a single record",
                "operationId": "getRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Records"
                ],
                "summary": "Delete a record",
                "operationId": "deleteRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Record deleted successfully"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Records"
                ],
                "summary": "Update a record",
                "operationId": "updateRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Record"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Record"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/records/search": {
            "get": {
                "tags": [
                    "Records"
                ],
                "summary": "Search records",
                "description": "Search records by a query string",
                "operationId": "3a26ef4b46f57f4ec834f0179e9d86b6",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "The query string to search for",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Record"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. User has not input the query."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Record": {
                "required": [
                    "title",
                    "release_year",
                    "imdb_id"
                ],
                "properties": {
                    "title": {
                        "description": "The title of the record",
                        "type": "string",
                        "example": "A Record Title"
                    },
                    "release_year": {
                        "description": "The release year of the record",
                        "type": "integer",
                        "example": 2023
                    },
                    "imdb_id": {
                        "description": "The IMDB id of the record",
                        "type": "string",
                        "example": "tt0000000"
                    },
                    "images": {
                        "description": "Images of the record",
                        "type": "string",
                        "example": "http://example.com/image.jpg"
                    }
                },
                "type": "object"
            }
        }
    }
}